move getPlaylists to client redux actions
